gcp_add_docs/requirements.txt
text_generation==0.3.0
openai==0.27.8
tiktoken
python-multipart
pinecone-client
langchain
PyPDF2